# backend/docs/openapi.yaml
definitions:
...
  models.User:
    properties:
      created_at:
        description: Creation timestamp (UTC)
        format: date-time
        type: string
      email:
        description: Email address (unique)
        example: user@example.com
        format: email
        type: string
      email_verified:
        description: Whether the email has been verified
        example: false
        type: boolean
      founder_id:
        description: Related founder profile ID
        example: 1
        type: integer
      id:
        description: Unique user identifier
        example: 1
        type: integer
      image_url:
        description: Avatar URL
        example: https://cdn.example.com/avatars/1.png
        format: uri
        type: string
      investor_id:
        description: Related investor profile ID
        example: 2
        type: integer
      name:
        description: Display name
        example: Jane Doe
        type: string
      role:
        description: Role name
        enum:
        - admin
        - user
        - investor
        - founder
        example: user
        type: string
      updated_at:
        description: Update timestamp (UTC)
        format: date-time
        type: string
    type: object
...
  requests.UserCreateRequest:
    properties:
      email:
        description: Email address
        example: jane@doe.tld
        format: email
        type: string
      image_url:
        description: Optional avatar URL
        example: https://cdn.example.com/avatars/jane.png
        format: uri
        type: string
      name:
        description: Display name
        example: Jane
        type: string
      password:
        description: Initial password
        example: secret123
        type: string
      role:
        description: Role to assign
        enum:
        - admin
        - user
        - investor
        - founder
        example: admin
        type: string
    type: object
  requests.UserUpdateRequest:
    properties:
      email:
        description: New email address
        example: jane@newmail.tld
        format: email
        type: string
      image_url:
        description: New avatar URL
        example: https://cdn.example.com/avatars/jane2.png
        format: uri
        type: string
      name:
        description: New display name
        example: Jane Doe
        type: string
      password:
        description: New password
        example: newSecret123
        type: string
      role:
        description: New role
        enum:
        - admin
        - user
        - investor
        - founder
        example: user
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        description: Machine-readable error code
        enum:
        - invalid_params
        - invalid_sort
        - invalid_payload
        - internal_error
        - not_found
        - email_taken
        - invalid_token
        - invalid_credentials
        - no_fields
        - method_not_allowed
        - unauthorized
        example: internal_error
        type: string
      details:
        description: Optional granular details (validation errors, etc.)
      message:
        description: Human-readable error message
        example: failed to retrieve resource
        type: string
    type: object
...
  response.ImageURLResponse:
    properties:
      image_url:
        type: string
    type: object
...
  response.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  response.PageMeta:
    properties:
      has_next:
        description: Whether there is a next page
        example: true
        type: boolean
      has_prev:
        description: Whether there is a previous page
        example: false
        type: boolean
      page:
        description: Current page number (1-based)
        example: 1
        minimum: 1
        type: integer
      per_page:
        description: Page size (items per page)
        example: 20
        minimum: 1
        type: integer
      total:
        description: Total number of items available
        example: 123
        minimum: 0
        type: integer
    type: object
...
  response.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.User'
        type: array
      pagination:
        $ref: '#/definitions/response.PageMeta'
    type: object
  response.UserObjectResponse:
    properties:
      data:
        $ref: '#/definitions/models.User'
    type: object
info:
  contact: {}
paths:
...
  /admin/users:
    post:
      consumes:
      - application/json
      description: Creates a user (admin required).
      parameters:
      - description: User
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.UserObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /admin/users/{id}:
    delete:
      description: Deletes a user by ID (admin required).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates a user (admin required).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
...
  /me:
    get:
      description: Returns the current authenticated user's profile.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserObjectResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: My profile
      tags:
      - Users
...
  /users:
    get:
      description: Returns a paginated list of users.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: per_page
        type: integer
      - default: created_at
        description: Sort field
        enum:
        - id
        - email
        - name
        - role
        - created_at
        - updated_at
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: List users
      tags:
      - Users
  /users/{id}:
    get:
      description: Retrieves a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserObjectResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - Users
  /users/{id}/image:
    get:
      description: Returns a user's image URL.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ImageURLResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: User image URL
      tags:
      - Users
  /users/email/{email}:
    get:
      description: Retrieves a user by email.
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserObjectResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get user by email
      tags:
      - Users
swagger: "2.0"